Move ChordType to a separate file
