+++ src/chord/mod.rs
@@ -1,12 +1,13 @@
 #![allow(clippy::module_inception)]
 mod chord;
 mod chord_shape;
+mod chord_type;
 mod tuning;
 
 pub use self::chord::Chord;
-pub use self::chord::ChordType;
 pub use self::chord_shape::ChordShape;
 pub use self::chord_shape::ChordShapeSet;
+pub use self::chord_type::ChordType;
 pub use self::tuning::Tuning;
 
 use crate::STRING_COUNT;
